+++ packages/api-security-admin-users/src/plugins/GroupsStorageOperationsProvider.ts
@@ -1,0 +1,11 @@
+import { Plugin } from "@webiny/plugins";
+import { AdminUsersContext, GroupsStorageOperations } from "~/types";
+
+interface Params {
+    context: AdminUsersContext;
+}
+export abstract class GroupsStorageOperationsProvider extends Plugin {
+    public static readonly type = "sau.storageOperationsProvider.group";
+
+    public abstract provide(params: Params): Promise<GroupsStorageOperations>;
+}