+++ packages/app-form-builder/src/admin/plugins/editor/formFieldValidators/patternPlugins/upperCase.ts
@@ -1,11 +1,11 @@
-import { FbFormFieldPatternValidatorPlugin } from "../../../../../types";
+import { FbBuilderFormFieldPatternValidatorPlugin } from "~/types";
 
 export default {
-    type: "fb-form-field-validator-pattern",
+    type: "form-editor-field-validator-pattern",
     name: "form-editor-field-validator-pattern-upper-case",
     pattern: {
         message: "Only upper case characters are allowed.",
         name: "upperCase",
         label: "Upper case"
     }
-} as FbFormFieldPatternValidatorPlugin;
+} as FbBuilderFormFieldPatternValidatorPlugin;