+++ packages/app-form-builder/src/render/plugins/validators/patternPlugins/url.ts
@@ -1,3 +1,5 @@
+import { FbFormFieldPatternValidatorPlugin } from "~/types";
+
 export default {
     type: "fb-form-field-validator-pattern",
     name: "form-field-validator-pattern-url",
@@ -7,4 +9,4 @@ export default {
             "^(ftp|http|https):\\/\\/(\\w+:{0,1}\\w*@)?(\\S+)(:[0-9]+)?(\\/|\\/([\\w#!:.?+=&%@!\\-\\/]))?$",
         flags: "i"
     }
-};
+} as FbFormFieldPatternValidatorPlugin;